move miner types
